% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_gui.R
\name{add_action}
\alias{add_action}
\title{Add action to history}
\usage{
add_action(rv, action)
}
\arguments{
\item{rv}{Reactive values object}

\item{action}{Action to add}
}
\description{
Add action to history
}
\keyword{internal}
