% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_gui.R, R/utils.R
\name{apply_action}
\alias{apply_action}
\title{Apply or reverse an action}
\usage{
apply_action(rv, action, reverse = FALSE)

apply_action(rv, action, reverse = FALSE)
}
\arguments{
\item{rv}{ReactiveValues object containing application state}

\item{action}{List containing action information}

\item{reverse}{Logical indicating whether to reverse the action}
}
\value{
Invisible rv (ReactiveValues object)

Invisible rv (ReactiveValues object)
}
\description{
Applies or reverses an action in the undo/redo system. Handles different types of
actions including adding/removing annotations and merging/unmerging codes.

Applies or reverses an action in the undo/redo system. Handles different types of
actions including adding/removing annotations and merging/unmerging codes.
}
\keyword{internal}
