% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_gui.R
\name{assign_by_proximity}
\alias{assign_by_proximity}
\title{Assign annotations by proximity}
\usage{
assign_by_proximity(annotations, max_distance)
}
\arguments{
\item{annotations}{Data frame of annotations}

\item{max_distance}{Maximum distance between annotations to be in same unit}
}
\value{
Data frame of unit assignments
}
\description{
Groups annotations into units based on their proximity to each other.
}
\keyword{internal}
