% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_gui.R
\name{calculate_hierarchy_stats}
\alias{calculate_hierarchy_stats}
\title{Calculate hierarchy statistics}
\usage{
calculate_hierarchy_stats(node)
}
\arguments{
\item{node}{Root node of the hierarchy tree}
}
\value{
A list containing hierarchy statistics:
\itemize{
\item total_themes: Total number of themes in the hierarchy
\item total_codes: Total number of codes in the hierarchy
\item max_depth: Maximum depth of the hierarchy tree
\item codes_per_theme: List showing number of codes in each theme
\item average_codes_per_theme: Average number of codes per theme
}
}
\description{
Calculates various statistics about the code hierarchy including the total number
of themes and codes, maximum depth, and distribution of codes across themes.
}
\keyword{internal}
