% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_gui.R, R/utils.R
\name{concatenate_memos}
\alias{concatenate_memos}
\title{Concatenate memo texts}
\usage{
concatenate_memos(existing_memo, new_memo)

concatenate_memos(existing_memo, new_memo)
}
\arguments{
\item{existing_memo}{Character string containing current memo text}

\item{new_memo}{Character string containing memo text to append}
}
\value{
Character string of combined memo text

Character string of combined memo text
}
\description{
Combines existing and new memo texts with proper separators,
handling empty memos appropriately.

Combines existing and new memo texts with proper separators,
handling empty memos appropriately.
}
\keyword{internal}
