% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_gui.R, R/project.R
\name{create_new_project}
\alias{create_new_project}
\title{Initialize new project}
\usage{
create_new_project(rv, session)

create_new_project(rv, session)
}
\arguments{
\item{rv}{ReactiveValues object to reset containing:
\itemize{
\item text: Text content
\item annotations: Annotation data frame
\item codes: Vector of codes
\item code_tree: Hierarchy Node object
\item All other project state values
}}

\item{session}{Shiny session object}
}
\value{
Invisible NULL, called for side effect

Invisible NULL, called for side effect
}
\description{
Creates new project by resetting all reactive values to defaults
and clearing UI elements.

Creates new project by resetting all reactive values to defaults
and clearing UI elements.
}
\keyword{internal}
