% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_gui.R
\name{get_all_codes_in_hierarchy}
\alias{get_all_codes_in_hierarchy}
\title{Get all codes in hierarchy}
\usage{
get_all_codes_in_hierarchy(node)

get_all_codes_in_hierarchy(node)
}
\arguments{
\item{node}{Root node of the hierarchy tree}
}
\value{
Character vector of all codes in the hierarchy

Character vector of all codes in the hierarchy
}
\description{
Recursively extracts all code names from the hierarchy tree.

Recursively extracts all code names from the hierarchy tree.
}
\keyword{internal}
