% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_calendar_icon}
\alias{get_calendar_icon}
\title{Get HTML entity for calendar icon}
\usage{
get_calendar_icon()
}
\value{
Character string containing HTML entity
}
\description{
Returns the HTML entity code for a calendar icon
}
\keyword{internal}
