% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_gui.R
\name{get_code_names}
\alias{get_code_names}
\title{Get all code names from hierarchy}
\usage{
get_code_names(node)
}
\arguments{
\item{node}{Root node of the code hierarchy (data.tree Node object)}
}
\value{
Character vector containing all code names in the hierarchy
}
\description{
Recursively extracts all code names from a code hierarchy tree structure,
traversing through all nodes and collecting their names.
}
\keyword{internal}
