% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_file_icon}
\alias{get_file_icon}
\title{Get HTML entity for file icon}
\usage{
get_file_icon()
}
\value{
Character string containing HTML entity
}
\description{
Returns the HTML entity code for a file/document icon
}
\keyword{internal}
