% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_folder_icon}
\alias{get_folder_icon}
\title{Get HTML entity for folder icon}
\usage{
get_folder_icon()
}
\value{
Character string containing HTML entity
}
\description{
Returns the HTML entity code for an open folder icon
}
\keyword{internal}
