% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_gui.R, R/directory_management.R,
%   R/project.R
\name{get_project_dir}
\alias{get_project_dir}
\title{Get project directory path}
\usage{
get_project_dir()

get_project_dir()

get_project_dir()
}
\value{
Character string containing the project directory path, or NULL if creation fails

Character string containing the project directory path, or NULL if creation fails

Character string containing the project directory path, or NULL if creation fails
}
\description{
Retrieves or creates the project directory path where all project files will be stored.
Creates the directory if it doesn't exist.

Retrieves or creates the project directory path where all project files will be stored.
Creates the directory if it doesn't exist.

Retrieves or creates the project directory path where all project files will be stored.
Creates the directory if it doesn't exist.
}
\keyword{internal}
