% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_gui.R
\name{getVolumes}
\alias{getVolumes}
\title{Get available storage volumes on Windows}
\usage{
getVolumes()
}
\value{
Function that returns named character vector of available drives
}
\description{
Creates a closure that returns a named vector of available Windows drive letters
and their corresponding paths. Checks for the existence of drives from A: to Z:
(excluding C: which is handled separately).
}
\keyword{internal}
