% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_gui.R, R/utils.R
\name{grapes-or-or-grapes}
\alias{grapes-or-or-grapes}
\alias{\%||\%}
\title{Null coalescing operator}
\usage{
a \%||\% b

a \%||\% b
}
\arguments{
\item{a}{First value to check}

\item{b}{Second value (default) to use if first is NULL}
}
\value{
Returns \code{a} if not NULL, otherwise returns \code{b}

Returns \code{a} if not NULL, otherwise returns \code{b}
}
\description{
Provides null coalescing functionality, returning the first non-NULL argument

Provides null coalescing functionality, returning the first non-NULL argument
}
\details{
\\%||\\% operator

\\%||\\% operator
}
\keyword{internal}
