% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_gui.R
\name{handle_dir_confirmation}
\alias{handle_dir_confirmation}
\title{Handle directory creation confirmation}
\usage{
handle_dir_confirmation(input, rv, session)
}
\arguments{
\item{input}{Shiny input object}

\item{rv}{ReactiveValues object containing application state}

\item{session}{Shiny session object}
}
\description{
Creates the data directory after receiving user confirmation
}
\keyword{internal}
