% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory_management.R
\name{handle_load_project_confirmation}
\alias{handle_load_project_confirmation}
\title{Enhanced project loading confirmation handler}
\usage{
handle_load_project_confirmation(input, output, session, rv, roots)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}

\item{rv}{ReactiveValues object containing application state}

\item{roots}{List of root directories for selection}
}
\description{
Handles the loading of projects with support for different storage modes
}
\keyword{internal}
