% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory_management.R
\name{handle_save_project_confirmation}
\alias{handle_save_project_confirmation}
\title{Enhanced save project confirmation handler}
\usage{
handle_save_project_confirmation(input, output, session, rv, roots)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}

\item{rv}{ReactiveValues object containing application state}

\item{roots}{List of root directories for selection}
}
\description{
Processes the project save confirmation based on storage mode
}
\keyword{internal}
