% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory_management.R
\name{handle_storage_confirmation}
\alias{handle_storage_confirmation}
\title{Handle directory storage confirmation}
\usage{
handle_storage_confirmation(input, output, rv, session, roots)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{rv}{ReactiveValues object containing application state}

\item{session}{Shiny session object}

\item{roots}{List of root directories for selection}
}
\description{
Processes the user's storage location choice and creates directories as needed
}
\keyword{internal}
