% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_gui.R
\name{is_ancestor}
\alias{is_ancestor}
\title{Check if one node is an ancestor of another}
\usage{
is_ancestor(potential_ancestor, node)
}
\arguments{
\item{potential_ancestor}{Node object to check as potential ancestor}

\item{node}{Node object to check ancestry against}
}
\value{
Logical indicating whether potential_ancestor is an ancestor of node
}
\description{
Helper function to check if a node is an ancestor of another node
in the code hierarchy, preventing circular references when moving items.
}
\keyword{internal}
