% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_gui.R
\name{load_project_from_path}
\alias{load_project_from_path}
\title{Load project from filepath}
\usage{
load_project_from_path(rv, filepath, session)
}
\arguments{
\item{rv}{ReactiveValues object containing application state}

\item{filepath}{Path to the project file}

\item{session}{Shiny session object}
}
\description{
Loads a project from a specified filepath
}
\keyword{internal}
