% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_gui.R, R/project.R
\name{load_project_interactive}
\alias{load_project_interactive}
\title{Display interactive project load dialog}
\usage{
load_project_interactive(rv, input, session, roots)

load_project_interactive(rv, input, session, roots)
}
\arguments{
\item{rv}{ReactiveValues object for project state}

\item{input}{Shiny input values}

\item{session}{Shiny session object}

\item{roots}{List of root directories for file selection}
}
\value{
Invisible NULL, called for side effect

Invisible NULL, called for side effect
}
\description{
Shows modal dialog for loading project with file selection functionality.

Shows modal dialog for loading project with file selection functionality.
}
\keyword{internal}
