% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_gui.R
\name{load_selected_project}
\alias{load_selected_project}
\title{Load selected project state}
\usage{
load_selected_project(rv, input, session, project_name)
}
\arguments{
\item{rv}{Reactive values object to update}

\item{input}{Shiny input object}

\item{session}{The current Shiny session}
}
\value{
Invisible NULL
}
\description{
Loads a previously saved project state and updates all reactive values.
}
\keyword{internal}
