% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_gui.R
\name{remove_annotation}
\alias{remove_annotation}
\title{Remove annotation without deleting code}
\usage{
remove_annotation(rv, start, end, code)
}
\arguments{
\item{rv}{ReactiveValues object containing annotations}

\item{start}{Starting position of annotation to remove}

\item{end}{Ending position of annotation to remove}

\item{code}{Code of annotation to remove}
}
\value{
Updated rv object with annotation removed
}
\description{
Removes a specific annotation instance while preserving the code
for use with other annotations.
}
\keyword{internal}
