% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_gui.R
\name{restore_node}
\alias{restore_node}
\title{Restore a node and its children in the hierarchy}
\usage{
restore_node(parent, node_data)
}
\arguments{
\item{parent}{Parent Node object where the node will be restored}

\item{node_data}{List containing node data to restore:
\itemize{
\item name: Character string of node name
\item type: Character string specifying node type
\item description: Character string of node description
\item created: POSIXct creation timestamp
\item children: List of child nodes
}}
}
\value{
New Node object with restored data and children
}
\description{
Helper function to recursively restore a node and all its children
when moving items in the code hierarchy.
}
\keyword{internal}
