% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_gui.R
\name{save_annotated_text_interactive}
\alias{save_annotated_text_interactive}
\title{Display interactive dialog for saving annotated text}
\usage{
save_annotated_text_interactive(rv, input, session, volumes)
}
\arguments{
\item{rv}{ReactiveValues object containing the application state}

\item{input}{Shiny input object}

\item{session}{Shiny session object}

\item{volumes}{List of available storage volumes for directory selection}
}
\value{
Invisible NULL, called for side effects
}
\description{
Creates and displays a modal dialog that allows users to save their annotated text
in either HTML or plain text format. Provides options for filename and directory selection.
}
\keyword{internal}
