% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_gui.R
\name{save_records_interactive}
\alias{save_records_interactive}
\title{Display interactive dialog for saving annotation records}
\usage{
save_records_interactive(rv, input, session, volumes)
}
\arguments{
\item{rv}{ReactiveValues object containing the application state}

\item{input}{Shiny input object}

\item{session}{Shiny session object}

\item{volumes}{List of available storage volumes for directory selection}
}
\value{
Invisible NULL, called for side effects
}
\description{
Creates and displays a modal dialog that allows users to save their annotation records
in either CSV or JSON format. Provides options for filename and directory selection.
}
\keyword{internal}
