% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{textAnnotatoR-package}
\alias{textAnnotatoR-package}
\title{textAnnotatoR: Interactive Text Annotation Tool with 'shiny' GUI}
\description{
A comprehensive text annotation tool built with 'shiny'. Provides an interactive graphical user interface for coding text documents, managing code hierarchies, creating memos, and analyzing coding patterns. Features include code co-occurrence analysis, visualization of coding patterns, comparison of multiple coding sets, and export capabilities. Supports collaborative qualitative research through standardized annotation formats and analysis tools.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/chaoliu-cl/textAnnotatoR}
  \item \url{http://liu-chao.site/textAnnotatoR/}
  \item Report bugs at \url{https://github.com/chaoliu-cl/textAnnotatoR/issues}
}

}
\author{
\strong{Maintainer}: Chao Liu \email{chaoliu@cedarville.edu} (\href{https://orcid.org/0000-0002-9979-8272}{ORCID}) [copyright holder]

}
\keyword{internal}
