% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_gui.R, R/textAnnotatoR-package.R
\docType{package}
\name{textAnnotatoR}
\alias{textAnnotatoR}
\title{Text Annotation Tool for R}
\description{
A Shiny application for interactive text annotation and analysis

Provides an interactive graphical user interface for qualitative text analysis,
allowing users to annotate text, manage codes, create memos, and visualize
coding patterns.
}
\details{
The package includes tools for:
\itemize{
\item Text annotation and coding
\item Code management and organization
\item Memo creation and linking
\item Project management
\item Visualization of coding patterns
\item Export of annotations and analyses
}
}
\note{
This package provides functionality for users to interactively save files
through the Shiny interface. All file operations are explicitly initiated by
users through file dialogs, and no files are written automatically to the user's
system without their direct action and consent.
}
\section{Dependencies}{

This package requires the following packages:
\itemize{
\item shiny
\item data.tree
\item jsonlite
\item shinydashboard
\item DT
\item readtext
\item magrittr
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/chaoliu-cl/textAnnotatoR}
  \item \url{http://liu-chao.site/textAnnotatoR/}
  \item Report bugs at \url{https://github.com/chaoliu-cl/textAnnotatoR/issues}
}

}
\author{
Chao Liu
}
\keyword{internal}
