% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_gui.R
\name{calculate_co_occurrences}
\alias{calculate_co_occurrences}
\title{Helper function to calculate co-occurrences}
\usage{
calculate_co_occurrences(annotations)
}
\arguments{
\item{annotations}{Data frame of annotations}
}
\value{
Table of co-occurrence frequencies
}
\description{
Helper function to calculate co-occurrences
}
