% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_gui.R
\name{calculate_transitions}
\alias{calculate_transitions}
\title{Helper function to calculate transitions}
\usage{
calculate_transitions(annotations)
}
\arguments{
\item{annotations}{Data frame of annotations}
}
\value{
List of transitions
}
\description{
Helper function to calculate transitions
}
