% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_gui.R
\name{compare_patterns}
\alias{compare_patterns}
\title{Compare coding patterns between different coders}
\usage{
compare_patterns(coding_strategies)
}
\description{
Compare coding patterns between different coders
}
\keyword{internal}
