% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_gui.R
\name{load_selected_project}
\alias{load_selected_project}
\title{Load project from a selected recent project}
\usage{
load_selected_project(rv, input, session)
}
\arguments{
\item{rv}{ReactiveValues object containing application state}

\item{input}{Shiny input object}

\item{session}{Shiny session object}
}
\description{
Loads a project from the list of recent projects
}
\keyword{internal}
