% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_gui.R, R/project.R
\name{save_project_interactive}
\alias{save_project_interactive}
\title{Display interactive project save dialog}
\usage{
save_project_interactive(rv, input, session)

save_project_interactive(rv, input, session)
}
\arguments{
\item{rv}{ReactiveValues object containing project state}

\item{input}{Shiny input values}

\item{session}{Shiny session object}
}
\value{
Invisible NULL, called for side effect

Invisible NULL, called for side effect
}
\description{
Shows modal dialog for saving project with directory selection and
project name input.

Shows modal dialog for saving project. If project was previously saved,
uses the same location unless user chooses to change it.
}
\keyword{internal}
